+++ frontend/src/components/FlowWidget.jsx
@@ -1,6 +1,11 @@
 import { useEffect, useState, useCallback, useMemo } from "react";
 import { useAuth } from "../hooks/useAuth";
 import { useTimeline } from "../hooks/useTimeline";
+import { useConceptGeneration } from "../hooks/useConceptGeneration";
+import { useScriptGeneration } from "../hooks/useScriptGeneration";
+import { useSegmentCreation } from "../hooks/useSegmentCreation";
+import { useImageGeneration } from "../hooks/useImageGeneration";
+import { useVideoGeneration } from "../hooks/useVideoGeneration";
 import ChatLoginButton from "./ChatLoginButton";
 import { projectApi } from "../services/project";
 import {
@@ -29,11 +34,6 @@ import UserProfileDropdown from "./UserProfileDropdown";
 
 // Assets and services
 import { assets } from "../assets/assets";
-import { webInfoApi } from "../services/web-info";
-import { conceptWriterApi } from "../services/concept-writer";
-import { segmentationApi } from "../services/segmentationapi";
-import { chatApi } from "../services/chat";
-import { s3Api } from "../services/s3";
 
 function FlowWidget() {
   const { isAuthenticated, user } = useAuth();
@@ -135,9 +135,56 @@ function FlowWidget() {
 
   const [projectName, setProjectName] = useState("Untitled");
 
-
-
   const [disableAutoGeneration, setDisableAutoGeneration] = useState(false);
+
+  // Initialize generation hooks
+  const { generateConcepts } = useConceptGeneration({
+    setNodes,
+    setEdges,
+    setGeneratingConcepts,
+    setUserConcepts,
+    setTaskCompletionStates,
+    saveGenerationState,
+    removeGenerationState,
+    nodes
+  });
+
+  const { generateScript } = useScriptGeneration({
+    setNodes,
+    setEdges,
+    setGeneratingScripts,
+    setTaskCompletionStates,
+    saveGenerationState,
+    removeGenerationState
+  });
+
+  const { createSegments } = useSegmentCreation({
+    setNodes,
+    setEdges,
+    setTaskCompletionStates
+  });
+
+  const { generateImage, regenerateImage } = useImageGeneration({
+    setNodes,
+    setEdges,
+    setGeneratingImages,
+    setTaskCompletionStates,
+    saveGenerationState,
+    removeGenerationState,
+    edges
+  });
+
+  const { generateVideo, regenerateVideo } = useVideoGeneration({
+    setNodes,
+    setEdges,
+    setGeneratingVideos,
+    setTaskCompletionStates,
+    setTemporaryVideos,
+    saveGenerationState,
+    removeGenerationState,
+    edges,
+    nodes
+  });
 
   // Restore generation states on component load
   const restoreGenerationStates = useCallback(async () => {
@@ -241,10 +288,6 @@ function FlowWidget() {
       console.error('Error restoring generation states:', error);
     }
   }, [setNodes, setEdges, setGeneratingConcepts, setGeneratingScripts, setGeneratingImages, setGeneratingVideos]);
-
-
-
-
 
   // Load data from API
   const flowData = useMemo(() => {
@@ -874,256 +917,8 @@ function FlowWidget() {
     );
     
     if (nodeType === 'userNode') {
-      // Get selected project from localStorage (following ChatWidget pattern)
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        console.error('No project selected');
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === nodeId) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                error: 'No project selected. Please select a project first.',
-                nodeState: 'error'
-              }
-            };
-          }
-          return node;
-        }));
-        return;
-      }
-      
-      // Store user node data in localStorage with project ID and user text
-      const userNodeDataKey = `userNodeData-${selectedProject.id}`;
-      const existingUserNodeData = JSON.parse(localStorage.getItem(userNodeDataKey) || '{}');
-      existingUserNodeData[nodeId] = {
-        projectId: selectedProject.id,
-        text: message
-      };
-      localStorage.setItem(userNodeDataKey, JSON.stringify(existingUserNodeData));
-
-
-      
-      // Update state
-      setUserConcepts(prev => new Map(prev.set(nodeId, message)));
-      
-      // Mark user input as completed
-      setTaskCompletionStates(prev => ({ ...prev, userInput: true }));
-      
-      // Update the user node to show user text immediately
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === nodeId) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              userText: message,
-              nodeState: 'user', // Mark as user input
-              content: message
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Create a single loading concept node first
-      const loadingConceptNodeId = `loading-concept-${nodeId}-${Date.now()}`;
-      const userNodePosition = nodes.find(n => n.id === nodeId)?.position || { x: 0, y: 0 };
-      const loadingConceptNode = {
-        id: loadingConceptNodeId,
-        type: 'conceptNode',
-        position: {
-          x: userNodePosition.x,
-          y: userNodePosition.y + 300
-        },
-        data: {
-          id: loadingConceptNodeId,
-          content: 'Generating concepts...',
-          nodeState: 'loading',
-          title: 'Loading Concepts'
-        }
-      };
-
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'concept', loadingConceptNodeId, {
-        loadingMessage: 'Generating concepts...',
-        title: 'Loading Concepts',
-        position: { x: userNodePosition.x, y: userNodePosition.y + 300 },
-        parentNodeId: nodeId,
-        originalPrompt: message
-      });
-      
-      // Create edge connecting user node to loading concept node
-      const loadingEdge = {
-        id: `${nodeId}-to-${loadingConceptNodeId}`,
-        source: nodeId,
-        target: loadingConceptNodeId,
-        sourceHandle: 'output',
-        targetHandle: 'input',
-        style: {
-          stroke: '#E9E8EB33',
-          strokeWidth: 2,
-          filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-        }
-      };
-      
-      // Add loading concept node and edge immediately
-      setNodes(prevNodes => [...prevNodes, loadingConceptNode]);
-      setEdges(prevEdges => [...prevEdges, loadingEdge]);
-
-      // Generate concepts from backend (following ChatWidget pattern)
-      try {
-        setGeneratingConcepts(prev => new Set(prev.add(loadingConceptNodeId)));
-        
-
-        
-        // First call web-info API (following ChatWidget pattern)
-        const webInfoResult = await webInfoApi.processWebInfo(
-          message,
-          selectedProject.id,
-        );
-        const webInfoContent = webInfoResult.choices[0].message.content;
-        const conceptsResponse = await conceptWriterApi.generateConcepts(
-          message,
-          webInfoContent,
-          selectedProject.id,
-        );
-        
-        if (conceptsResponse && conceptsResponse.concepts && Array.isArray(conceptsResponse.concepts)) {
-          // Remove persistent generation state and clean up localStorage
-          removeGenerationState(selectedProject.id, 'concept', loadingConceptNodeId);
-          
-          // Mark user data as processed
-          try {
-            const userNodeDataKey = `userNodeData-${selectedProject.id}`;
-            const existingUserNodeData = JSON.parse(localStorage.getItem(userNodeDataKey) || '{}');
-            Object.keys(existingUserNodeData).forEach(key => {
-              if (existingUserNodeData[key].text === message) {
-                existingUserNodeData[key].processed = true;
-                existingUserNodeData[key].processedAt = Date.now();
-              }
-            });
-            localStorage.setItem(userNodeDataKey, JSON.stringify(existingUserNodeData));
-          } catch (error) {
-            console.error('Error updating user node data:', error);
-          }
-          
-          // Remove the loading concept node and its edge
-          setNodes(prevNodes => prevNodes.filter(node => node.id !== loadingConceptNodeId));
-          setEdges(prevEdges => prevEdges.filter(edge => edge.target !== loadingConceptNodeId));
-          
-          // Create concept nodes from backend response
-          const userNodePosition = nodes.find(n => n.id === nodeId)?.position || { x: 0, y: 0 };
-          const conceptCount = conceptsResponse.concepts.length;
-          const conceptSpacing = 350; // Space between concepts
-          const totalWidth = (conceptCount - 1) * conceptSpacing;
-          const startX = userNodePosition.x - (totalWidth / 2); // Center the concepts under user node
-          const timestamp = Date.now(); // Use single timestamp for consistency
-          
-          // Extract concept IDs for mapping
-          const conceptIds = [];
-          
-          const newConceptNodes = conceptsResponse.concepts.map((concept, index) => {
-            const conceptNodeId = `generated-concept-${nodeId}-${index}-${timestamp}`;
-            const conceptId = concept.id || `concept-${index}-${timestamp}`;
-            
-            // Store concept ID for mapping
-            conceptIds.push(conceptId);
-            
-            return {
-              id: conceptNodeId,
-              type: 'conceptNode',
-              position: {
-                x: startX + index * conceptSpacing,
-                y: userNodePosition.y + 300
-              },
-              data: {
-                ...concept,
-                nodeState: 'existing',
-                content: concept.content || concept.text || concept.concept || concept.description || concept.title,
-                title: concept.title || `Generated Concept ${index + 1}`,
-                id: conceptId,
-                conceptId: conceptId // Store the actual concept ID from backend
-              }
-            };
-          });
-          
-
-          
-          // Create edges connecting user node to generated concepts
-          const newEdges = conceptsResponse.concepts.map((_, index) => {
-            const conceptNodeId = `generated-concept-${nodeId}-${index}-${timestamp}`;
-            return {
-              id: `${nodeId}-to-${conceptNodeId}`,
-              source: nodeId,
-              target: conceptNodeId,
-              sourceHandle: 'output',
-              targetHandle: 'input',
-              style: {
-                stroke: '#E9E8EB33',
-                strokeWidth: 2,
-                filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-              }
-            };
-          });
-          
-          // Add new nodes and edges
-          setNodes(prevNodes => [...prevNodes, ...newConceptNodes]);
-          setEdges(prevEdges => [...prevEdges, ...newEdges]);
-          
-
-          
-          setTaskCompletionStates(prev => ({ ...prev, concept: true }));
-        } else {
-          throw new Error('Invalid response format from concept generation API');
-        }
-      } catch (error) {
-        console.error('Error generating concepts:', error);
-        // Save error state instead of removing it immediately
-        saveGenerationState(selectedProject.id, 'concept', loadingConceptNodeId, {
-          loadingMessage: 'Failed to generate concepts',
-          title: 'Error',
-          position: { x: userNodePosition.x, y: userNodePosition.y + 300 },
-          parentNodeId: nodeId,
-          originalPrompt: message,
-          status: 'error',
-          errorMessage: error.message || 'Failed to generate concepts'
-        });
-        
-        // Show error in the loading concept node
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === loadingConceptNodeId) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                content: 'Failed to generate concepts',
-                error: error.message || 'Failed to generate concepts',
-                nodeState: 'error'
-              }
-            };
-          }
-          return node;
-        }));
-        
-          setTimeout(() => {
-          removeGenerationState(selectedProject.id, 'concept', loadingConceptNodeId);
-        }, 5000);
-      } finally {
-        setGeneratingConcepts(prev => {
-          const newSet = new Set(prev);
-          newSet.delete(loadingConceptNodeId);
-          return newSet;
-        });
-      }
+      // Use the concept generation hook
+      await generateConcepts(message, nodeId);
     }
   }, [nodes, setNodes, setEdges, setError]);
 
@@ -1325,1170 +1120,42 @@ function FlowWidget() {
       if (sourceNode && targetNode) {
         // Script connected to concept - generate script
         if (sourceNode.type === 'conceptNode' && targetNode.type === 'scriptNode' && targetNode.data?.nodeState === 'new') {
-          await handleScriptGeneration(sourceNode, targetNode);
+          await generateScript(sourceNode, targetNode);
         }
         
         // Segment connected to script - create 5 segment nodes
         if (sourceNode.type === 'scriptNode' && targetNode.type === 'segmentNode' && targetNode.data?.nodeState === 'new') {
-          await handleSegmentCreation(sourceNode, targetNode);
+          await createSegments(sourceNode, targetNode);
         }
         
         // Image connected to segment - generate image
         if (sourceNode.type === 'segmentNode' && targetNode.type === 'imageNode' && targetNode.data?.nodeState === 'new') {
-          await handleImageGeneration(sourceNode, targetNode);
+          await generateImage(sourceNode, targetNode);
         }
         
         // Video connected to image - generate video
         if (sourceNode.type === 'imageNode' && targetNode.type === 'videoNode' && targetNode.data?.nodeState === 'new') {
-          await handleVideoGeneration(sourceNode, targetNode);
+          await generateVideo(sourceNode, targetNode);
         }
       }
     },
-    [setEdges, nodes, disableAutoGeneration],
+    [setEdges, nodes, disableAutoGeneration, generateScript, createSegments, generateImage, generateVideo],
   );
   
-  // Handle script generation from concept
-  const handleScriptGeneration = useCallback(async (conceptNode, scriptNode) => {
-    try {
-      setGeneratingScripts(prev => new Set(prev.add(scriptNode.id)));
-      
-      // Update script node to show loading
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === scriptNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Generating scripts...',
-              nodeState: 'loading'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Get selected project from localStorage (following ChatWidget pattern)
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        throw new Error('No project selected. Please select a project first.');
-      }
-      
-      // Update script node to show loading
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === scriptNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Generating script...',
-              nodeState: 'loading'
-            }
-          };
-        }
-        return node;
-      }));
-
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'script', scriptNode.id, {
-        loadingMessage: 'Generating script...',
-        title: 'Loading Script',
-        position: scriptNode.position,
-        parentNodeId: conceptNode.id,
-        conceptContent: conceptNode.data?.content || conceptNode.data?.userText || ''
-      });
-      
-      // Generate 2 different scripts using concept content
-      const conceptContent = conceptNode.data?.content || conceptNode.data?.userText || '';
-      const conceptTitle = conceptNode.data?.title || 'Selected Concept';
-      
-      console.log("Starting script generation for concept:", conceptTitle);
-      
-      // Generate 2 different scripts by making 2 API calls with different approaches
-      const [scriptResponse1, scriptResponse2] = await Promise.all([
-        segmentationApi.getSegmentation({
-          prompt: `Generate a detailed cinematic script for: ${conceptContent}`,
-          concept: `${conceptTitle} - Cinematic Style`,
-          negative_prompt: "",
-          project_id: selectedProject.id,
-          model: 'flash'
-        }),
-        segmentationApi.getSegmentation({
-          prompt: `Generate an alternative creative script for: ${conceptContent}`,
-          concept: `${conceptTitle} - Creative Style`,
-          negative_prompt: "",
-          project_id: selectedProject.id,
-          model: 'flash'
-        })
-      ]);
-      
-      console.log("Script generation responses:", { scriptResponse1, scriptResponse2 });
-      
-      if (scriptResponse1 && scriptResponse1.segments && Array.isArray(scriptResponse1.segments) &&
-          scriptResponse2 && scriptResponse2.segments && Array.isArray(scriptResponse2.segments)) {
-        
-        // Remove persistent generation state and clean up
-        removeGenerationState(selectedProject.id, 'script', scriptNode.id);
-        
-        // Clean up any temporary script generation data
-        try {
-          const tempScriptKey = `temp-script-${selectedProject.id}-${scriptNode.id}`;
-          localStorage.removeItem(tempScriptKey);
-          console.log('✅ Cleaned up temporary script data from localStorage');
-        } catch (error) {
-          console.error('Error cleaning up temporary script data:', error);
-        }
-        
-        // Remove the placeholder script node
-        setNodes(prevNodes => prevNodes.filter(node => node.id !== scriptNode.id));
-        setEdges(prevEdges => prevEdges.filter(edge => edge.target !== scriptNode.id));
-        
-        // Create 2 new script nodes
-        const newScriptNodes = [];
-        const newEdges = [];
-        
-        // Create first script node (Cinematic)
-        const script1NodeId = `script-${conceptNode.id}-1-${Date.now()}`;
-        const script1Id = scriptResponse1.id || `script-cinematic-${Date.now()}`;
-        newScriptNodes.push({
-          id: script1NodeId,
-          type: 'scriptNode',
-          position: {
-            x: scriptNode.position.x - 200,
-            y: scriptNode.position.y
-          },
-          data: {
-            id: script1NodeId,
-            scriptId: script1Id, // Store the actual script ID from backend
-            content: `Cinematic Script - ${scriptResponse1.segments.length} segments`,
-            segments: scriptResponse1.segments,
-            nodeState: 'existing',
-            title: `Cinematic Script`,
-            artStyle: scriptResponse1.artStyle || 'cinematic photography with soft lighting',
-            concept: conceptTitle
-          }
-        });
-        
-        // Create second script node (Creative)
-        const script2NodeId = `script-${conceptNode.id}-2-${Date.now()}`;
-        const script2Id = scriptResponse2.id || `script-creative-${Date.now()}`;
-        newScriptNodes.push({
-          id: script2NodeId,
-          type: 'scriptNode',
-          position: {
-            x: scriptNode.position.x + 200,
-            y: scriptNode.position.y
-          },
-          data: {
-            id: script2NodeId,
-            scriptId: script2Id, // Store the actual script ID from backend
-            content: `Creative Script - ${scriptResponse2.segments.length} segments`,
-            segments: scriptResponse2.segments,
-            nodeState: 'existing',
-            title: `Creative Script`,
-            artStyle: scriptResponse2.artStyle || 'creative artistic style',
-            concept: conceptTitle
-          }
-        });
-        
-        // Create edges connecting concept to both script nodes
-        newEdges.push({
-          id: `${conceptNode.id}-to-${script1NodeId}`,
-          source: conceptNode.id,
-          target: script1NodeId,
-          sourceHandle: 'output',
-          targetHandle: 'input',
-          style: {
-            stroke: '#E9E8EB33',
-            strokeWidth: 2,
-            filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-          }
-        });
-        
-        newEdges.push({
-          id: `${conceptNode.id}-to-${script2NodeId}`,
-          source: conceptNode.id,
-          target: script2NodeId,
-          sourceHandle: 'output',
-          targetHandle: 'input',
-          style: {
-            stroke: '#E9E8EB33',
-            strokeWidth: 2,
-            filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-          }
-        });
-        
-        // Add new nodes and edges
-        setNodes(prevNodes => [...prevNodes, ...newScriptNodes]);
-        setEdges(prevEdges => [...prevEdges, ...newEdges]);
-        
-
-        
-        // Mark script generation as completed
-        setTaskCompletionStates(prev => ({ ...prev, script: true }));
-        
-        console.log(`Generated 2 different script nodes from concept`);
-      } else {
-        throw new Error('Invalid response format from script generation API');
-      }
-    } catch (error) {
-      console.error('Error generating script:', error);
-      // Save error state instead of removing it immediately
-      saveGenerationState(selectedProject.id, 'script', scriptNode.id, {
-        loadingMessage: 'Failed to generate script',
-        title: 'Error',
-        position: scriptNode.position,
-        parentNodeId: conceptNode.id,
-        conceptContent: conceptNode.data?.content || conceptNode.data?.userText || '',
-        status: 'error',
-        errorMessage: error.message || 'Failed to generate script'
-      });
-      
-      // Show error in script node
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === scriptNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Failed to generate script',
-              error: error.message || 'Failed to generate script',
-              nodeState: 'error'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Clean up error state after 5 seconds
-      setTimeout(() => {
-        removeGenerationState(selectedProject.id, 'script', scriptNode.id);
-      }, 5000);
-    } finally {
-      setGeneratingScripts(prev => {
-        const newSet = new Set(prev);
-        newSet.delete(scriptNode.id);
-        return newSet;
-      });
-    }
-  }, [setNodes, setError]);
   
-  // Handle automatic segment creation
-  const handleSegmentCreation = useCallback(async (scriptNode, segmentNode) => {
-    const segments = scriptNode.data?.segments || [];
-    
-    if (segments.length === 0) {
-      console.error('No segments found in script node');
-      return;
-    }
-    
-    // Remove the placeholder segment node
-    setNodes(prevNodes => prevNodes.filter(node => node.id !== segmentNode.id));
-    setEdges(prevEdges => prevEdges.filter(edge => edge.target !== segmentNode.id));
-    
-    // Create segment nodes from script segments (limit to 5)
-    const segmentsToCreate = segments.slice(0, 5);
-    const timestamp = Date.now(); // Generate timestamp once for consistent IDs
-    
-    const segmentIds = []; // Track segment IDs for mapping
-    const newSegmentNodes = segmentsToCreate.map((segment, index) => {
-      const segmentNodeId = `segment-${scriptNode.id}-${index}-${timestamp}`;
-      const segmentId = segment.segmentId || segment.id || `segment-${index + 1}-${timestamp}`;
-      
-      // Store segment ID for mapping
-      segmentIds.push(segmentId);
-      
-      return {
-        id: segmentNodeId,
-        type: 'segmentNode',
-        position: {
-          x: scriptNode.position.x + (index - 2) * 420,
-          y: scriptNode.position.y + 400
-        },
-        data: {
-          ...segment,
-          id: segmentId,
-          segmentId: segmentId, // Store the actual segment ID from backend
-          nodeState: 'existing',
-          title: `Segment ${index + 1}`,
-          visual: segment.visual || '',
-          narration: segment.narration || '',
-          animation: segment.animation || ''
-        }
-      };
-    });
-    
-    // Create edges connecting script to segments
-    const newEdges = segmentsToCreate.map((_, index) => {
-      const segmentNodeId = `segment-${scriptNode.id}-${index}-${timestamp}`;
-      return {
-        id: `${scriptNode.id}-to-${segmentNodeId}`,
-        source: scriptNode.id,
-        target: segmentNodeId,
-        sourceHandle: 'output',
-        targetHandle: 'input',
-        style: {
-          stroke: '#E9E8EB33',
-          strokeWidth: 2,
-          filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-        }
-      };
-    });
-    
-    // Add new nodes and edges
-    setNodes(prevNodes => [...prevNodes, ...newSegmentNodes]);
-    setEdges(prevEdges => [...prevEdges, ...newEdges]);
-    
-
-    
-    // Mark segment creation as completed
-    setTaskCompletionStates(prev => ({ ...prev, segment: true }));
-  }, [setNodes, setEdges]);
-  
-  // Handle image generation from segment
-  const handleImageGeneration = useCallback(async (segmentNode, imageNode) => {
-    try {
-      setGeneratingImages(prev => new Set(prev.add(imageNode.id)));
-      
-      // Get selected project from localStorage (following ChatWidget pattern)
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        throw new Error('No project selected. Please select a project first.');
-      }
-      
-      // Update image node to show loading
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === imageNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Generating image...',
-              nodeState: 'loading'
-            }
-          };
-        }
-        return node;
-      }));
-
-      // Generate image using segment visual prompt (following ChatWidget pattern)
-      const visualPrompt = segmentNode.data?.visual || 'A cinematic scene';
-      const artStyle = segmentNode.data?.artStyle || 'cinematic photography with soft lighting';
-      const segmentId = segmentNode.data?.id || Date.now();
-      
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'image', imageNode.id, {
-        loadingMessage: 'Generating image...',
-        title: 'Loading Image',
-        position: imageNode.position,
-        parentNodeId: segmentNode.id,
-        visualPrompt: visualPrompt,
-        artStyle: artStyle,
-        segmentId: segmentId
-      });
-      
-      console.log("Starting image generation...");
-      console.log("Image generation request:", {
-        visual_prompt: visualPrompt,
-        art_style: artStyle,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: 'imagen' // Using default model like ChatWidget
-      });
-      
-      const imageResponse = await chatApi.generateImage({
-        visual_prompt: visualPrompt,
-        art_style: artStyle,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: 'imagen' // Using default model like ChatWidget
-      });
-      
-      console.log("Image generation response:", imageResponse);
-      
-      if (imageResponse && imageResponse.s3_key) {
-        // Remove persistent generation state and clean up
-        removeGenerationState(selectedProject.id, 'image', imageNode.id);
-        
-        // Clean up any temporary image generation data
-        try {
-          const tempImageKey = `temp-image-${selectedProject.id}-${imageNode.id}`;
-          localStorage.removeItem(tempImageKey);
-          console.log('✅ Cleaned up temporary image data from localStorage');
-        } catch (error) {
-          console.error('Error cleaning up temporary image data:', error);
-        }
-        
-        // Download image URL (following ChatWidget pattern)
-        const imageUrl = await s3Api.downloadImage(imageResponse.s3_key);
-        
-        // Update image node with generated image
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === imageNode.id) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                imageUrl: imageUrl,
-                imageId: imageResponse.id,
-                s3Key: imageResponse.s3_key,
-                visualPrompt: visualPrompt,
-                artStyle: artStyle,
-                nodeState: 'existing',
-                segmentId: segmentId,
-                segmentData: segmentNode.data
-              }
-            };
-          }
-          return node;
-        }));
-        
-
-        
-        console.log(`Generated image for segment ${segmentId}`);
-        
-        // Mark image generation as completed
-        setTaskCompletionStates(prev => ({ ...prev, image: true }));
-      } else {
-        throw new Error('No image key returned from API');
-      }
-    } catch (error) {
-      console.error('Error generating image:', error);
-      // Save error state instead of removing it immediately
-      saveGenerationState(selectedProject.id, 'image', imageNode.id, {
-        loadingMessage: 'Failed to generate image',
-        title: 'Error',
-        position: imageNode.position,
-        parentNodeId: segmentNode.id,
-        visualPrompt: visualPrompt,
-        artStyle: artStyle,
-        segmentId: segmentId,
-        status: 'error',
-        errorMessage: error.message || 'Failed to generate image'
-      });
-      
-      // Show error in image node
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === imageNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Failed to generate image',
-              error: error.message || 'Failed to generate image',
-              nodeState: 'error'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Clean up error state after 5 seconds
-      setTimeout(() => {
-        removeGenerationState(selectedProject.id, 'image', imageNode.id);
-      }, 5000);
-    } finally {
-      setGeneratingImages(prev => {
-        const newSet = new Set(prev);
-        newSet.delete(imageNode.id);
-        return newSet;
-      });
-    }
-  }, [setNodes, setError]);
-  
-  // Handle image regeneration from sidebar
-  const handleImageRegeneration = useCallback(async (regenerationParams) => {
-    try {
-      const { selectedNode, prompt, model, artStyle } = regenerationParams;
-      
-      if (!selectedNode || selectedNode.type !== 'imageNode') {
-        console.error('Invalid node for image regeneration:', selectedNode);
-        return;
-      }
-
-      // Get selected project from localStorage
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        throw new Error('No project selected. Please select a project first.');
-      }
-
-      // Create a new image node for the regenerated image
-      const regeneratedImageNodeId = `regenerated-image-${selectedNode.id}-${Date.now()}`;
-      const segmentId = selectedNode.data?.segmentId || Date.now();
-      
-      // Position the new node next to the original image node
-      const newPosition = {
-        x: selectedNode.position.x + 350, // Place it to the right
-        y: selectedNode.position.y
-      };
-
-      // Create loading image node
-      const loadingImageNode = {
-        id: regeneratedImageNodeId,
-        type: 'imageNode',
-        position: newPosition,
-        data: {
-          id: regeneratedImageNodeId,
-          content: 'Regenerating image...',
-          nodeState: 'loading',
-          title: 'Regenerating Image',
-          visualPrompt: prompt,
-          artStyle: artStyle,
-          segmentId: segmentId,
-          segmentData: selectedNode.data?.segmentData
-        }
-      };
-
-      // Add the loading node immediately
-      setNodes(prevNodes => [...prevNodes, loadingImageNode]);
-
-      // Connect the new node to the same parent as the original image node
-      const parentEdge = edges.find(edge => edge.target === selectedNode.id);
-      if (parentEdge) {
-        const newEdge = {
-          id: `${parentEdge.source}-to-${regeneratedImageNodeId}`,
-          source: parentEdge.source,
-          target: regeneratedImageNodeId,
-          sourceHandle: 'output',
-          targetHandle: 'input',
-          style: {
-            stroke: '#E9E8EB33',
-            strokeWidth: 2,
-            filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-          }
-        };
-        setEdges(prevEdges => [...prevEdges, newEdge]);
-      }
-
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'image', regeneratedImageNodeId, {
-        loadingMessage: 'Regenerating image...',
-        title: 'Regenerating Image',
-        position: newPosition,
-        parentNodeId: parentEdge?.source,
-        visualPrompt: prompt,
-        artStyle: artStyle,
-        segmentId: segmentId
-      });
-
-      // Mark as generating
-      setGeneratingImages(prev => new Set(prev.add(regeneratedImageNodeId)));
-
-      console.log("Starting image regeneration...");
-      console.log("Image regeneration request:", {
-        visual_prompt: prompt,
-        art_style: artStyle,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: model
-      });
-
-      // Generate the image using the chat API
-      const imageResponse = await chatApi.generateImage({
-        visual_prompt: prompt,
-        art_style: artStyle,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: model
-      });
-
-      console.log("Image regeneration response:", imageResponse);
-
-      if (imageResponse && imageResponse.s3_key) {
-        // Remove persistent generation state
-        removeGenerationState(selectedProject.id, 'image', regeneratedImageNodeId);
-
-        // Download image URL
-        const imageUrl = await s3Api.downloadImage(imageResponse.s3_key);
-
-        // Update the loading node with the generated image
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === regeneratedImageNodeId) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                imageUrl: imageUrl,
-                imageId: imageResponse.id,
-                s3Key: imageResponse.s3_key,
-                visualPrompt: prompt,
-                artStyle: artStyle,
-                nodeState: 'existing',
-                segmentId: segmentId,
-                segmentData: selectedNode.data?.segmentData,
-                title: 'Regenerated Image'
-              }
-            };
-          }
-          return node;
-        }));
-
-        console.log(`Successfully regenerated image for segment ${segmentId}`);
-      } else {
-        throw new Error('No image key returned from API');
-      }
-    } catch (error) {
-      console.error('Error regenerating image:', error);
-      
-      // Save error state for the regenerated image node
-      if (selectedProject) {
-        saveGenerationState(selectedProject.id, 'image', regeneratedImageNodeId, {
-          loadingMessage: 'Failed to regenerate image',
-          title: 'Error',
-          position: newPosition,
-          parentNodeId: parentEdge?.source,
-          visualPrompt: prompt,
-          artStyle: artStyle,
-          segmentId: segmentId,
-          status: 'error',
-          errorMessage: error.message || 'Failed to regenerate image'
-        });
-      }
-      
-      // Show error in the loading node if it exists
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.data?.nodeState === 'loading' && node.data?.title === 'Regenerating Image') {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Failed to regenerate image',
-              error: error.message || 'Failed to regenerate image',
-              nodeState: 'error'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Clean up error state after 5 seconds
-      if (selectedProject) {
-        setTimeout(() => {
-          removeGenerationState(selectedProject.id, 'image', regeneratedImageNodeId);
-        }, 5000);
-      }
-    } finally {
-      // Clean up generating state
-      setGeneratingImages(prev => {
-        const newSet = new Set(prev);
-        // Remove any loading nodes from generating set
-        nodes.forEach(node => {
-          if (node.data?.nodeState === 'loading' && node.data?.title === 'Regenerating Image') {
-            newSet.delete(node.id);
-          }
-        });
-        return newSet;
-      });
-    }
-  }, [nodes, edges, setNodes, setEdges, setGeneratingImages, saveGenerationState, removeGenerationState]);
-
-  // Handle video regeneration from sidebar
-  const handleVideoRegeneration = useCallback(async (regenerationParams) => {
-    try {
-      const { selectedNode, prompt, model, artStyle } = regenerationParams;
-      
-      if (!selectedNode || selectedNode.type !== 'videoNode') {
-        console.error('Invalid node for video regeneration:', selectedNode);
-        return;
-      }
-
-      // Get selected project from localStorage
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        throw new Error('No project selected. Please select a project first.');
-      }
-
-      // Create a new video node for the regenerated video
-      const regeneratedVideoNodeId = `regenerated-video-${selectedNode.id}-${Date.now()}`;
-      const segmentId = selectedNode.data?.segmentId || Date.now();
-      const imageId = selectedNode.data?.imageId || Date.now();
-      
-      // Position the new node next to the original video node
-      const newPosition = {
-        x: selectedNode.position.x + 350, // Place it to the right
-        y: selectedNode.position.y
-      };
-
-      // Create loading video node
-      const loadingVideoNode = {
-        id: regeneratedVideoNodeId,
-        type: 'videoNode',
-        position: newPosition,
-        data: {
-          id: regeneratedVideoNodeId,
-          content: 'Regenerating video...',
-          nodeState: 'loading',
-          title: 'Regenerating Video',
-          animationPrompt: prompt,
-          artStyle: artStyle,
-          segmentId: segmentId,
-          imageId: imageId,
-          segmentData: selectedNode.data?.segmentData
-        }
-      };
-
-      // Add the loading node immediately
-      setNodes(prevNodes => [...prevNodes, loadingVideoNode]);
-
-      // Connect the new node to the same parent as the original video node
-      const parentEdge = edges.find(edge => edge.target === selectedNode.id);
-      if (parentEdge) {
-        const newEdge = {
-          id: `${parentEdge.source}-to-${regeneratedVideoNodeId}`,
-          source: parentEdge.source,
-          target: regeneratedVideoNodeId,
-          sourceHandle: 'output',
-          targetHandle: 'input',
-          style: {
-            stroke: '#E9E8EB33',
-            strokeWidth: 2,
-            filter: 'drop-shadow(0 0 6px rgba(233, 232, 235, 0.2))'
-          }
-        };
-        setEdges(prevEdges => [...prevEdges, newEdge]);
-      }
-
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'video', regeneratedVideoNodeId, {
-        loadingMessage: 'Regenerating video...',
-        title: 'Regenerating Video',
-        position: newPosition,
-        parentNodeId: parentEdge?.source,
-        animationPrompt: prompt,
-        artStyle: artStyle,
-        segmentId: segmentId,
-        imageId: imageId
-      });
-
-      // Mark as generating
-      setGeneratingVideos(prev => new Set(prev.add(regeneratedVideoNodeId)));
-
-      // We need the parent image's S3 key for video generation
-      const parentImageNode = nodes.find(node => node.id === parentEdge?.source);
-      let imageS3Key = parentImageNode?.data?.s3Key || parentImageNode?.data?.imageS3Key;
-
-      // If no direct s3Key, try to extract from imageUrl
-      if (!imageS3Key && parentImageNode?.data?.imageUrl) {
-        const imageUrl = parentImageNode.data.imageUrl;
-        if (imageUrl.includes('cloudfront.net/')) {
-          const urlParts = imageUrl.split('cloudfront.net/');
-          if (urlParts.length > 1) {
-            imageS3Key = urlParts[1];
-          }
-        }
-      }
-
-      if (!imageS3Key) {
-        throw new Error('No parent image S3 key found for video regeneration. Video generation requires a source image.');
-      }
-
-      console.log("Starting video regeneration...");
-      console.log("Video regeneration request:", {
-        animation_prompt: prompt,
-        art_style: artStyle,
-        image_s3_key: imageS3Key,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: model
-      });
-
-      // Generate the video using the chat API
-      const videoResponse = await chatApi.generateVideo({
-        animation_prompt: prompt,
-        art_style: artStyle,
-        image_s3_key: imageS3Key,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: model
-      });
-
-      console.log("Video regeneration response:", videoResponse);
-
-      if (videoResponse && videoResponse.s3_key) {
-        // Remove persistent generation state
-        removeGenerationState(selectedProject.id, 'video', regeneratedVideoNodeId);
-
-        // Download video URL
-        const videoUrl = await s3Api.downloadVideo(videoResponse.s3_key);
-
-        // Update the loading node with the generated video
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === regeneratedVideoNodeId) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                videoUrl: videoUrl,
-                videoId: videoResponse.id,
-                segmentId: segmentId,
-                imageId: imageId,
-                animationPrompt: prompt,
-                artStyle: artStyle,
-                nodeState: 'existing',
-                segmentData: selectedNode.data?.segmentData,
-                title: 'Regenerated Video'
-              }
-            };
-          }
-          return node;
-        }));
-
-        // Store temporary video for immediate display
-        setTemporaryVideos(prev => {
-          const newMap = new Map(prev);
-          const key = `${segmentId}-${imageId}`;
-          newMap.set(key, videoUrl);
-          return newMap;
-        });
-
-        // Save video to localStorage for persistence across refreshes
-        try {
-          const videoStorageKey = `generated-videos-${selectedProject.id}`;
-          const existingVideos = JSON.parse(localStorage.getItem(videoStorageKey) || '{}');
-          existingVideos[`${segmentId}-${imageId}`] = {
-            videoUrl: videoUrl,
-            videoId: videoResponse.id,
-            segmentId: segmentId,
-            imageId: imageId,
-            animationPrompt: prompt,
-            artStyle: artStyle,
-            model: model,
-            timestamp: Date.now()
-          };
-          localStorage.setItem(videoStorageKey, JSON.stringify(existingVideos));
-          console.log(`💾 Saved regenerated video to localStorage for ${segmentId}-${imageId}`);
-          
-          // Clean up old videos (keep only last 50 videos per project)
-          const videoEntries = Object.entries(existingVideos);
-          if (videoEntries.length > 50) {
-            const sortedVideos = videoEntries.sort((a, b) => b[1].timestamp - a[1].timestamp);
-            const videosToKeep = sortedVideos.slice(0, 50);
-            const cleanedVideos = Object.fromEntries(videosToKeep);
-            localStorage.setItem(videoStorageKey, JSON.stringify(cleanedVideos));
-            console.log(`🧹 Cleaned up old videos, kept ${videosToKeep.length} most recent`);
-          }
-        } catch (error) {
-          console.error('Error saving video to localStorage:', error);
-        }
-
-        console.log(`Successfully regenerated video for segment ${segmentId}`);
-      } else {
-        throw new Error('No video key returned from API');
-      }
-    } catch (error) {
-      console.error('Error regenerating video:', error);
-      
-      // Save error state for the regenerated video node
-      if (selectedProject) {
-        saveGenerationState(selectedProject.id, 'video', regeneratedVideoNodeId, {
-          loadingMessage: 'Failed to regenerate video',
-          title: 'Error',
-          position: newPosition,
-          parentNodeId: parentEdge?.source,
-          animationPrompt: prompt,
-          artStyle: artStyle,
-          segmentId: segmentId,
-          imageId: imageId,
-          status: 'error',
-          errorMessage: error.message || 'Failed to regenerate video'
-        });
-      }
-      
-      // Show error in the loading node if it exists
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.data?.nodeState === 'loading' && node.data?.title === 'Regenerating Video') {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Failed to regenerate video',
-              error: error.message || 'Failed to regenerate video',
-              nodeState: 'error'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Clean up error state after 5 seconds
-      if (selectedProject) {
-        setTimeout(() => {
-          removeGenerationState(selectedProject.id, 'video', regeneratedVideoNodeId);
-        }, 5000);
-      }
-    } finally {
-      // Clean up generating state
-      setGeneratingVideos(prev => {
-        const newSet = new Set(prev);
-        // Remove any loading nodes from generating set
-        nodes.forEach(node => {
-          if (node.data?.nodeState === 'loading' && node.data?.title === 'Regenerating Video') {
-            newSet.delete(node.id);
-          }
-        });
-        return newSet;
-      });
-    }
-  }, [nodes, edges, setNodes, setEdges, setGeneratingVideos, setTemporaryVideos, saveGenerationState, removeGenerationState]);
-
-  // Unified regeneration handler that routes to image or video regeneration
   const handleRegeneration = useCallback(async (regenerationParams) => {
     const { nodeType } = regenerationParams;
     
     if (nodeType === 'image') {
-      await handleImageRegeneration(regenerationParams);
+      await regenerateImage(regenerationParams);
     } else if (nodeType === 'video') {
-      await handleVideoRegeneration(regenerationParams);
+      await regenerateVideo(regenerationParams);
     } else {
       console.error('Unsupported node type for regeneration:', nodeType);
     }
-  }, [handleImageRegeneration, handleVideoRegeneration]);
+  }, [regenerateImage, regenerateVideo]);
 
-  // Handle video generation from image
-  const handleVideoGeneration = useCallback(async (imageNode, videoNode) => {
-    try {
-      setGeneratingVideos(prev => new Set(prev.add(videoNode.id)));
-      
-      // Get selected project from localStorage (following ChatWidget pattern)
-      let selectedProject = null;
-      try {
-        const storedProject = localStorage.getItem('project-store-selectedProject');
-        selectedProject = storedProject ? JSON.parse(storedProject) : null;
-      } catch (e) {
-        console.error('Error parsing project data:', e);
-      }
-      
-      if (!selectedProject) {
-        throw new Error('No project selected. Please select a project first.');
-      }
-      
-      // Update video node to show loading
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === videoNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Generating video...',
-              nodeState: 'loading'
-            }
-          };
-        }
-        return node;
-      }));
 
-      // Generate video using image and segment animation prompt (following ChatWidget pattern)
-      const animationPrompt = imageNode.data?.segmentData?.animation || imageNode.data?.segmentData?.visual || 'Smooth cinematic movement';
-      const artStyle = imageNode.data?.artStyle || 'cinematic photography with soft lighting';
-      const segmentId = imageNode.data?.segmentId || Date.now();
-      // Try multiple possible s3Key field names for backward compatibility
-      let imageS3Key = imageNode.data?.s3Key || imageNode.data?.imageS3Key || imageNode.data?.image_s3_key;
-
-      // Save persistent generation state
-      saveGenerationState(selectedProject.id, 'video', videoNode.id, {
-        loadingMessage: 'Generating video...',
-        title: 'Loading Video',
-        position: videoNode.position,
-        parentNodeId: imageNode.id,
-        animationPrompt: animationPrompt,
-        artStyle: artStyle,
-        segmentId: segmentId,
-        imageId: imageNode.data?.imageId
-      });
-      
-      // If no direct s3Key, try to extract from imageUrl
-      if (!imageS3Key && imageNode.data?.imageUrl) {
-        const imageUrl = imageNode.data.imageUrl;
-        if (imageUrl.includes('cloudfront.net/')) {
-          const urlParts = imageUrl.split('cloudfront.net/');
-          if (urlParts.length > 1) {
-            imageS3Key = urlParts[1];
-            console.log('Extracted s3Key from imageUrl:', imageS3Key);
-          }
-        }
-      }
-      
-      if (!imageS3Key) {
-        console.error('Image node data:', imageNode.data);
-        console.error('Available keys:', Object.keys(imageNode.data));
-        throw new Error('No image S3 key found for video generation. Check if the image has s3Key property or valid imageUrl.');
-      }
-      
-      console.log("Starting video generation...");
-      console.log("Video generation request:", {
-        animation_prompt: animationPrompt,
-        art_style: artStyle,
-        image_s3_key: imageS3Key,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: 'gen4_turbo' // Using RunwayML gen4_turbo model
-      });
-      
-      const videoResponse = await chatApi.generateVideo({
-        animation_prompt: animationPrompt,
-        art_style: artStyle,
-        image_s3_key: imageS3Key,
-        segmentId: segmentId,
-        project_id: selectedProject.id,
-        model: 'gen4_turbo' // Using RunwayML gen4_turbo model
-      });
-      
-      console.log("Video generation response:", videoResponse);
-      
-      if (videoResponse && videoResponse.s3_key) {
-        // Remove persistent generation state and clean up
-        removeGenerationState(selectedProject.id, 'video', videoNode.id);
-        
-        // Clean up any temporary video generation data
-        try {
-          const tempVideoKey = `temp-video-${selectedProject.id}-${videoNode.id}`;
-          localStorage.removeItem(tempVideoKey);
-          console.log('✅ Cleaned up temporary video data from localStorage');
-        } catch (error) {
-          console.error('Error cleaning up temporary video data:', error);
-        }
-        
-        // Download video URL (following ChatWidget pattern)
-        const videoUrl = await s3Api.downloadVideo(videoResponse.s3_key);
-        
-        // Update video node with generated video
-        setNodes(prevNodes => prevNodes.map(node => {
-          if (node.id === videoNode.id) {
-            return {
-              ...node,
-              data: {
-                ...node.data,
-                videoUrl: videoUrl,
-                videoId: videoResponse.id,
-                segmentId: segmentId,
-                imageId: imageNode.data?.imageId,
-                animationPrompt: animationPrompt,
-                nodeState: 'existing',
-                segmentData: imageNode.data?.segmentData
-              }
-            };
-          }
-          return node;
-        }));
-        
-
-        
-        // Store temporary video for immediate display
-        setTemporaryVideos(prev => {
-          const newMap = new Map(prev);
-          const key = `${segmentId}-${imageNode.data?.imageId}`;
-          newMap.set(key, videoUrl);
-          return newMap;
-        });
-
-        // Save video to localStorage for persistence across refreshes
-        try {
-          const videoStorageKey = `generated-videos-${selectedProject.id}`;
-          const existingVideos = JSON.parse(localStorage.getItem(videoStorageKey) || '{}');
-          existingVideos[`${segmentId}-${imageNode.data?.imageId}`] = {
-            videoUrl: videoUrl,
-            videoId: videoResponse.id,
-            segmentId: segmentId,
-            imageId: imageNode.data?.imageId,
-            animationPrompt: animationPrompt,
-            artStyle: artStyle,
-            model: 'gen4_turbo', // Fixed: use the actual model name instead of undefined variable
-            timestamp: Date.now()
-          };
-          localStorage.setItem(videoStorageKey, JSON.stringify(existingVideos));
-          console.log(`💾 Saved generated video to localStorage for ${segmentId}-${imageNode.data?.imageId}`);
-          
-          // Clean up old videos (keep only last 50 videos per project)
-          const videoEntries = Object.entries(existingVideos);
-          if (videoEntries.length > 50) {
-            const sortedVideos = videoEntries.sort((a, b) => b[1].timestamp - a[1].timestamp);
-            const videosToKeep = sortedVideos.slice(0, 50);
-            const cleanedVideos = Object.fromEntries(videosToKeep);
-            localStorage.setItem(videoStorageKey, JSON.stringify(cleanedVideos));
-            console.log(`🧹 Cleaned up old videos, kept ${videosToKeep.length} most recent`);
-          }
-        } catch (error) {
-          console.error('Error saving video to localStorage:', error);
-        }
-        
-        console.log(`Generated video for segment ${segmentId}`);
-        
-        // Mark video generation as completed
-        setTaskCompletionStates(prev => ({ ...prev, video: true }));
-      } else {
-        throw new Error('No video key returned from API');
-      }
-    } catch (error) {
-      console.error('Error generating video:', error);
-      // Save error state instead of removing it immediately
-      saveGenerationState(selectedProject.id, 'video', videoNode.id, {
-        loadingMessage: 'Failed to generate video',
-        title: 'Error',
-        position: videoNode.position,
-        parentNodeId: imageNode.id,
-        animationPrompt: animationPrompt,
-        artStyle: artStyle,
-        segmentId: segmentId,
-        imageId: imageNode.data?.imageId,
-        status: 'error',
-        errorMessage: error.message || 'Failed to generate video'
-      });
-      
-      // Show error in video node
-      setNodes(prevNodes => prevNodes.map(node => {
-        if (node.id === videoNode.id) {
-          return {
-            ...node,
-            data: {
-              ...node.data,
-              content: 'Failed to generate video',
-              error: error.message || 'Failed to generate video',
-              nodeState: 'error'
-            }
-          };
-        }
-        return node;
-      }));
-      
-      // Clean up error state after 5 seconds
-      setTimeout(() => {
-        removeGenerationState(selectedProject.id, 'video', videoNode.id);
-      }, 5000);
-    } finally {
-      setGeneratingVideos(prev => {
-        const newSet = new Set(prev);
-        newSet.delete(videoNode.id);
-        return newSet;
-      });
-    }
-  }, [setNodes, setTemporaryVideos, setError]);
